+++ source/raw/bin/x64/plugins/cyber_engine_tweaks/mods/Character Preset Manager (CET)/modules/ui/sections_save.lua
@@ -36,7 +36,7 @@ if collapsibleSectionHeader("SAVE & REPLACE PRESETS", "create") then
         end
       end
       ImGui.EndChild()
-      ImGui.Unindent(8)
+      finishCompactSubsection()
     end
     ImGui.Spacing()
     ImGui.PushItemWidth(-1)